stable_diffusion/requirements.txt
mlx
safetensors
huggingface-hub
regex
numpy
tqdm
Pillow
gradio
